refactor: merge svg plugin and merge config types
