--- packages/core/src/rspack-provider/types/config/dev.ts
@@ -1,8 +1,0 @@
-import type {
-  SharedDevConfig,
-  NormalizedSharedDevConfig,
-} from '@rsbuild/shared';
-
-export type DevConfig = SharedDevConfig;
-
-export type NormalizedDevConfig = NormalizedSharedDevConfig;
