--- packages/core/src/rspack-provider/types/config/html.ts
@@ -1,8 +1,0 @@
-import type {
-  SharedHtmlConfig,
-  NormalizedSharedHtmlConfig,
-} from '@rsbuild/shared';
-
-export type HtmlConfig = SharedHtmlConfig;
-
-export type NormalizedHtmlConfig = NormalizedSharedHtmlConfig;
