--- packages/core/src/rspack-provider/types/config/output.ts
@@ -1,15 +1,0 @@
-import type {
-  SharedOutputConfig,
-  NormalizedSharedOutputConfig,
-} from '@rsbuild/shared';
-import type { Builtins } from '@rspack/core';
-
-export type OutputConfig = SharedOutputConfig & {
-  /**
-   * Copies the specified file or directory to the dist directory.
-   */
-  copy?: Builtins['copy'] | NonNullable<Builtins['copy']>['patterns'];
-};
-
-export type NormalizedOutputConfig = OutputConfig &
-  NormalizedSharedOutputConfig;
