--- packages/core/src/rspack-provider/types/config/performance.ts
@@ -1,8 +1,0 @@
-import type {
-  SharedPerformanceConfig,
-  NormalizedSharedPerformanceConfig,
-} from '@rsbuild/shared';
-
-export type PerformanceConfig = SharedPerformanceConfig;
-
-export type NormalizedPerformanceConfig = NormalizedSharedPerformanceConfig;
