--- packages/core/src/rspack-provider/types/config/security.ts
@@ -1,8 +1,0 @@
-import type {
-  SharedSecurityConfig,
-  NormalizedSharedSecurityConfig,
-} from '@rsbuild/shared';
-
-export type SecurityConfig = SharedSecurityConfig;
-
-export type NormalizedSecurityConfig = NormalizedSharedSecurityConfig;
