--- packages/core/src/rspack-provider/types/config/source.ts
@@ -1,16 +1,0 @@
-import type {
-  SharedSourceConfig,
-  NormalizedSharedSourceConfig,
-} from '@rsbuild/shared';
-import type { Builtins } from '@rspack/core';
-import type { RspackBuiltinsConfig } from '../rspack';
-
-export type SourceConfig = SharedSourceConfig & {
-  define?: RspackBuiltinsConfig['define'];
-  transformImport?: false | Builtins['pluginImport'];
-};
-
-export type NormalizedSourceConfig = NormalizedSharedSourceConfig & {
-  define: Record<string, string>;
-  transformImport?: false | Builtins['pluginImport'];
-};
