+++ packages/core/tests/plugins/asset.test.ts
@@ -1,6 +1,14 @@
-import * as shared from '@rsbuild/shared';
+import { FONT_EXTENSIONS, IMAGE_EXTENSIONS } from '@rsbuild/shared';
 import { createStubRsbuild } from '@rsbuild/test-helper';
-import { pluginAsset } from '@src/plugins/asset';
+import { pluginAsset, getRegExpForExts } from '@src/plugins/asset';
+
+describe('getRegExpForExts', () => {
+  it('should get correct RegExp of exts', () => {
+    expect(getRegExpForExts(FONT_EXTENSIONS)).toEqual(
+      /\.(woff|woff2|eot|ttf|otf|ttc)$/i,
+    );
+  });
+});
 
 describe('plugin-asset(image)', () => {
   const cases = [
@@ -42,7 +50,7 @@ describe('plugin-asset(image)', () => {
 
   it.each(cases)('$name', async (item) => {
     const rsbuild = await createStubRsbuild({
-      plugins: [pluginAsset('image', shared.IMAGE_EXTENSIONS)],
+      plugins: [pluginAsset('image', IMAGE_EXTENSIONS)],
       rsbuildConfig: item.rsbuildConfig,
     });
