+++ packages/shared/tests/regexp.test.ts
@@ -1,10 +1,4 @@
-import {
-  JS_REGEX,
-  TS_REGEX,
-  mergeRegex,
-  getRegExpForExts,
-  FONT_EXTENSIONS,
-} from '../src';
+import { JS_REGEX, TS_REGEX, mergeRegex } from '../src';
 
 describe('mergeRegex', () => {
   it('should merge two regexp correctly', () => {
@@ -15,13 +9,5 @@ describe('mergeRegex', () => {
 
   it('should merge regexp and string correctly', () => {
     expect(mergeRegex(TS_REGEX, 'foo')).toEqual(/\.(ts|mts|cts|tsx)$|foo/);
-  });
-});
-
-describe('getRegExpForExts', () => {
-  it('should get correct RegExp of exts', () => {
-    expect(getRegExpForExts(FONT_EXTENSIONS)).toEqual(
-      /\.(woff|woff2|eot|ttf|otf|ttc)$/i,
-    );
   });
 });
